edulens-agents/requirements.txt
strands-agents>=0.3.0
strands-agents-tools>=0.1.0
bedrock-agentcore>=1.0.0
boto3>=1.38.0
psycopg2-binary>=2.9.0
